refactor: move scenario creation states
