+++ packages/integrations/vue-devtools/index.ts
@@ -4,7 +4,7 @@ import type { Exposition } from '@exposition/core'
 import type { DevtoolsContext, OnUpdateHandler } from './@types/api'
 import { defineDevtoolsSettings } from './functions/defineDevtoolsSettings'
 import { defineExpositionState } from './functions/defineExpositionState'
-import { updateState } from './functions/updateState'
+import { updateState } from './functions/utils'
 import { expositionLabel, id, stateType } from './utils/config'
 import { createInspectorViews } from './views/createInspectorViews'
 import { createSettingsViews } from './views/createSettingsViews'
