+++ packages/integrations/vue-devtools/utils/config.ts
@@ -6,3 +6,8 @@ export const stateType = 'My Awesome Plugin state'
 export const inspectorId = `${id}/inspector`
 export const timelineId = `${id}/timeline`
 export const expositionLabel = '📖 Exposition'
+
+export const warningLabelSettings = {
+  backgroundColor: 0xFBA02D,
+  textColor: 0x000000,
+}
