+++ packages/integrations/vue-devtools/views/createInspectorViews.ts
@@ -1,19 +1,15 @@
 import type { Exposition } from '@exposition/core'
 import type { CustomInspectorNode } from '@vue/devtools-api'
 import type { DevtoolsContext } from '../@types/api'
-import { updateState } from '../functions/updateState'
-import { inspectorId } from '../utils/config'
+import { createUpdateStateHandler } from '../functions/utils'
+
+import { inspectorId, warningLabelSettings } from '../utils/config'
 import { actionLog } from '../utils/logs'
+import { createMainScenarioView } from './inspector/createMainScenarioView'
+import { createScenarioDetailView } from './inspector/createScenarioDetailView'
 
 export function createInspectorViews<T extends Exposition<any>>(context: DevtoolsContext<T>) {
-  const { api, state } = context
-
-  function createUpdateStateHandler(beforeUpdateHandler: Function = () => undefined): () => void {
-    return function () {
-      beforeUpdateHandler()
-      updateState(context)
-    }
-  }
+  const { api, state, settings } = context
 
   api.addInspector({
     id: inspectorId,
@@ -26,7 +22,7 @@ export function createInspectorViews<T extends Exposition<any>>(context: Devtool
         action: createUpdateStateHandler(() => {
           actionLog('clicked restore action')
           state.reset()
-        }),
+        }, context),
       },
     ],
   })
@@ -43,15 +39,12 @@ export function createInspectorViews<T extends Exposition<any>>(context: Devtool
           {
             id,
             label: id,
-            tags: [
-              isInitialValue
-                ? {
-                    label: 'default',
-                    textColor: 0xFFFFFF,
-                    backgroundColor: 0x000000,
-                  }
-                : { label: 'modified', textColor: 0xFFC495, backgroundColor: 0xAF4C05 },
-            ],
+            tags: !isInitialValue
+              ? [
+
+                  { ...warningLabelSettings, label: 'modified' },
+                ]
+              : [],
           },
         ]
       }, [] as CustomInspectorNode[])
@@ -60,13 +53,14 @@ export function createInspectorViews<T extends Exposition<any>>(context: Devtool
         {
           id: 'settings',
           label: '⚙ Settings',
-          tags: [
-            {
-              label: 'awesome',
-              textColor: 0xFFFFFF,
-              backgroundColor: 0x000000,
-            },
-          ],
+          tags: !settings.isEnabled('active')
+            ? [
+                {
+                  ...warningLabelSettings,
+                  label: 'mocking inactive',
+                },
+              ]
+            : [],
         },
         {
           id: 'scenarios',
@@ -77,84 +71,6 @@ export function createInspectorViews<T extends Exposition<any>>(context: Devtool
     }
   })
 
-  api.on.getInspectorState((payload) => {
-    if (payload.inspectorId === inspectorId) {
-      if (payload.nodeId === 'scenarios') {
-        payload.state = {
-          // @ts-expect-error - figure out why TypeScript is angry
-          scenarios: Object.keys(state.getValues()).reduce((accumulator, key) => {
-            return [
-              ...accumulator,
-              {
-                key,
-                value: state.value[key].value,
-              },
-            ]
-          }, []),
-        }
-      }
-
-      if (Object.keys(state.value).includes(payload.nodeId)) {
-        const scenario = state.value[payload.nodeId]
-
-        payload.state = {
-          state: [
-            {
-              key: 'scenario',
-              value: scenario.id,
-            },
-            {
-              key: 'initialValue',
-              value: scenario.initialValue,
-            },
-            {
-              key: 'value',
-              value: {
-                _custom: {
-                  type: null,
-                  value: scenario.value,
-                  actions: [{
-                    icon: 'restore',
-                    tooltip: 'Reset the value of the scenario',
-                    action: createUpdateStateHandler(() => {
-                      actionLog('restore scenario %s settings', scenario.id)
-                      // @ts-expect-error - Allow dynamic state definition in this case
-                      state.update({
-                        [scenario.id]: scenario.initialValue,
-                      })
-                    }),
-                  }],
-                },
-              },
-              editable: false,
-            },
-          ],
-          // @ts-expect-error - figure out why TypeScript is angry
-          options: scenario.options.map((option, index) => {
-            return {
-              key: index,
-              value: {
-                _custom: {
-                  type: null,
-                  value: option,
-                  actions: [{
-                    icon: 'check',
-                    tooltip: `Set "${option}" as the new value`,
-                    action: createUpdateStateHandler(() => {
-                      actionLog('set new value "%s" for scenario "%s"', option, scenario.id)
-                      // @ts-expect-error - Allow dynamic state definition in this case
-                      state.update({
-                        [scenario.id]: option,
-                      })
-                    }),
-                  }],
-                },
-              },
-              editable: false,
-            }
-          }),
-        }
-      }
-    }
-  })
+  createMainScenarioView(context)
+  createScenarioDetailView(context)
 }
